// src/server/app/models/interfaces/hitting.interface.ts
export interface Hitting {
    BB: string;
    AB: string;
    battingOrder: string;
    IBB: string;
    H: string;
    HR: string;
    substitutionOrder: string;
    TB: string;
    ['3B']: string;
    GIDP: string;
    ['2B']: string;
    R: string;
    SF: string;
    SAC: string;
    HBP: string;
    RBI: string;
    SO: string;
    AVG: string;
}